split alert docs demos into individual files, delete demo app, migrate E2E to docs site
