--- demo/index.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <meta http-equiv="Cache-control" content="no-cache" />
-        <meta http-equiv="Expires" content="-1" />
-        <title>Vue-Foundation Demo</title>
-    </head>
-    <body>
-        <div id="app"></div>
-        <script type="module" src="/index.ts"></script>
-    </body>
-</html>
