+++ docs/e2e/alerts.md
@@ -1,0 +1,9 @@
+---
+layout: false
+---
+
+<E2EAlertHarness />
+
+<script setup>
+import E2EAlertHarness from '../demos/e2e/E2EAlertHarness.vue';
+</script>
