+++ docs/e2e/smart-select.md
@@ -1,0 +1,9 @@
+---
+layout: false
+---
+
+<E2ESmartSelectHarness />
+
+<script setup>
+import E2ESmartSelectHarness from '../demos/e2e/E2ESmartSelectHarness.vue';
+</script>
