+++ tsconfig.app.json
@@ -1,6 +1,6 @@
 {
     "extends": "@vue/tsconfig/tsconfig.dom.json",
-    "include": ["src/**/*.ts", "src/**/*.vue", "demo/**/*.ts", "demo/**/*.vue"],
+    "include": ["src/**/*.ts", "src/**/*.vue"],
     "exclude": ["src/**/__tests__/*", "src/vite-plugins/*"],
     "compilerOptions": {
         "noEmit": false,
